#48 Rename submodule (dataframe → frame)
